Update backstage.yaml to fix errors in backstage
The backstage system gets alot of errors because the backstage.yaml file is missing parts. This tries to fix those issues.

However, the spec.definition and spec.type of the API are required, but we do not have the spec/definition for them. So not sure if these will work.